requirements.txt
beautifulsoup4==4.13.5
Flask==3.1.2
pymongo==4.14.1
Requests==2.32.5
google-generativeai==0.8.3
